requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
openai>=1.30.0
Pillow==10.4.0
fastapi==0.114.2
uvicorn==0.30.6